Difference of Squares

DifferenceOfSquares.square_of_sum — Function

square_of_sum(n::Integer)::Integer

Square the sum of the first n positive integers.

\[\begin{align*} \left(\sum_{x=1}^n x\right)^2 &= \left(\frac{n(n + 1)}{2}\right)^2 \\ &= \frac{n^2 + 2n^3 + n^4}{4} \end{align*}\]

DifferenceOfSquares.sum_of_squares — Function

sum_of_squares(n::Integer)::Integer

Sum the squares of the first n positive integers.

\[\begin{align*} \sum_{x=1}^n x^2 &= \frac{n(n + 1)(2n + 1)}{6} \\ &= \frac{n + 3n^2 + 2n^3}{6} \end{align*}\]

DifferenceOfSquares.difference — Function

difference(n::Integer)::Integer

Subtract the sum of squares from square of the sum of the first n positive integers.

\[\begin{align*} \left(\sum_{x=1}^n x\right)^2 - \sum_{x=1}^n x^2 &= \left(\frac{n(n + 1)}{2}\right)^2 - \frac{n(n + 1)(2n + 1)}{6} \\ &= \frac{n^2 + 2n^3 + n^4}{4} - \frac{n + 3n^2 + 2n^3}{6} \\ &= \frac{-2n - 3n^2 + 2n^3 + 3n^4}{12} \end{align*}\]